requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
influxdb-client==1.32.0
msgpack==1.0.4
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
reactivex==4.0.4
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
xmltodict==0.13.0
